AI_Service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
google-search-results
sentence-transformers